requirements.txt
Flask==3.1.1
Flask-CORS==6.0.1
gspread==6.2.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
pandas==2.2.3
pypdf==5.7.0
